src/org/launchcode/techjobs_oo/Job.java
package org.launchcode.techjobs_oo;

import java.util.Objects;

public class Job {

    private int id;
    private static int nextId = 1;

    private String name;
    private Employer employer;
    private Location location;
    private PositionType positionType;
    private CoreCompetency coreCompetency;

    // TODO: Add two constructors - one to initialize a unique ID and a second to initialize the
    //  other five fields. The second constructor should also call the first in order to initialize
    //  the 'id' field.

    public Job() {
        id = nextId;
        nextId++;
    }

    public Job(String name, Employer employer, Location location, PositionType positionType, CoreCompetency coreCompetency){
        this();
        this.name = name;
        this.employer = employer;
        this.location = location;
        this.positionType = positionType;
        this.coreCompetency = coreCompetency;
    }


    public static String toString(Job jobInfo){

        int jobInfoID = jobInfo.getId();
        String jobInfoName = jobInfo.getName();
        String jobInfoEmployer = jobInfo.getEmployer().toString();
        String jobInfoLocation = jobInfo.getLocation().toString();
        String jobInfoPositionType = jobInfo.getPositionType().toString();
        String jobInfoCoreCompetency = jobInfo.getCoreCompetency().toString();

        if(jobInfoName==""){
            jobInfoName="Data not available";
        }
        if(jobInfoEmployer==""){
            jobInfoEmployer="Data not available";
        }
        if(jobInfoLocation==""){
            jobInfoLocation="Data not available";
        }
        if(jobInfoPositionType==""){
            jobInfoPositionType="Data not available";
        }
        if(jobInfoCoreCompetency==""){
            jobInfoCoreCompetency="Data not available";
        }

        return (" \nID: " + jobInfoID + "\nName: " + jobInfoName + "\nEmployer: " + jobInfoEmployer + "\nLocation: " + jobInfoLocation + "\nPosition Type: " + jobInfoPositionType + "\nCore Competency: "+ jobInfoCoreCompetency + "\n ");
    }
    // TODO: Add custom equals and hashCode methods. Consider two Job objects "equal" when their id fields
    //  match.

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Job job = (Job) o;
        return id == job.id &&
                Objects.equals(name, job.name) &&
                Objects.equals(employer, job.employer) &&
                Objects.equals(location, job.location) &&
                Objects.equals(positionType, job.positionType) &&
                Objects.equals(coreCompetency, job.coreCompetency);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, name, employer, location, positionType, coreCompetency);
    }


    // TODO: Add getters for each field EXCEPT nextId. Add setters for each field EXCEPT nextID
    //  and id.


    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Employer getEmployer() {
        return employer;
    }

    public void setEmployer(Employer employer) {
        this.employer = employer;
    }

    public Location getLocation() {
        return location;
    }

    public void setLocation(Location location) {
        this.location = location;
    }

    public PositionType getPositionType() {
        return positionType;
    }

    public void setPositionType(PositionType positionType) {
        this.positionType = positionType;
    }

    public CoreCompetency getCoreCompetency() {
        return coreCompetency;
    }

    public void setCoreCompetency(CoreCompetency coreCompetency) {
        this.coreCompetency = coreCompetency;
    }

    public int getId() { return id; }
}
